+++ servant-examples/tutorial/T8.hs
@@ -3,7 +3,7 @@
 {-# LANGUAGE TypeOperators #-}
 module T8 where
 
-import           Control.Monad.Trans.Either
+import           Control.Monad.Trans.Except
 import           Data.Aeson
 import           Servant
 import           Servant.Client
@@ -12,20 +12,20 @@ import           T3
 
 position :: Int -- ^ value for "x"
          -> Int -- ^ value for "y"
-         -> EitherT ServantError IO Position
+         -> ExceptT ServantError IO Position
 
 hello :: Maybe String -- ^ an optional value for "name"
-      -> EitherT ServantError IO HelloMessage
+      -> ExceptT ServantError IO HelloMessage
 
 marketing :: ClientInfo -- ^ value for the request body
-          -> EitherT ServantError IO Email
+          -> ExceptT ServantError IO Email
 
 position :<|> hello :<|> marketing = client api baseUrl
 
 baseUrl :: BaseUrl
 baseUrl = BaseUrl Http "localhost" 8081 ""
 
-queries :: EitherT ServantError IO (Position, HelloMessage, Email)
+queries :: ExceptT ServantError IO (Position, HelloMessage, Email)
 queries = do
   pos <- position 10 10
   msg <- hello (Just "servant")
@@ -34,7 +34,7 @@ queries = do
 
 run :: IO ()
 run = do
-  res <- runEitherT queries
+  res <- runExceptT queries
   case res of
     Left err -> putStrLn $ "Error: " ++ show err
     Right (pos, msg, em) -> do